requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-RESTx==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
argon2-cffi==23.1.0
passlib==1.7.4
werkzeug==2.3.7
requests==2.32.3